A version using 2d array instead of physics
To keep testing simple, only two sprites are added. I'll add more sprites only if the game mechanics work on these 2 sprites. In this version, movement and 2d array function correctly, and I'm moving forward to implement the game mechanics using the 2d array. Many codes can be reused, but for the prototypes I don't want to focus too much on the structure of the code.